==> Part 2.ipynb <==
import re
import requests
import urllib

def word_break(text_br):
    words = re.findall(r'\b\S+\b', text_br) 
    return words

def find_accession(query):
    global K
    K=K+1
    print (K, end="\r")
    ac_uniprot = ""
    id_uniprot = ""
    try:
        res = urllib.request.urlopen("http://www.uniprot.org/uniprot/" + query + ".txt").read()
        s = str(res, 'utf-8')
#         print(s)
#         print(res)
        match_word=word_break(s)
#         print(match_word)
        for i in range(len(match_word)):
#                 print(match_word[i])
                if match_word[i] == "AC":
                    ac_uniprot =  match_word[i+1] 
                    break;
        for i in range(len(match_word)):
                if match_word[i] == "ID":
                    id_uniprot =  match_word[i+1]
                    break;
        return ac_uniprot,id_uniprot
    except urllib.error.HTTPError:
  # Do something here to handle the error. For example:
      print("URL", query, "could not be read.")
      return ac_uniprot,id_uniprot


K=0
find_accession("P13595")

import pandas as pd
import numpy as np

data_2dge = pd.read_excel('2dge AC ID.xlsx')
data_HRMS = pd.read_excel('HRMS AC ID.xlsx')

data_2dge

data_HRMS

# data.apply(lambda row: check_equal(row), axis=1)

for index, row in data_2dge.iterrows() :
    id_given = row["Accession Name"]
    id_uniprot = row["ID_uniprot"]
    if id_given != id_uniprot:
        print(index,id_given,id_uniprot,row["AC_uniprot"])

# # data.apply(lambda row: check_equal(row), axis=1)

# for index, row in data_HRMS.iterrows() :
#     id_given = row["Accession Name"]
#     id_uniprot = row["ID_uniprot"]
#     if id_given != id_uniprot:
#         print(index,id_given,id_uniprot,row["AC_uniprot"])

data_2dge[data_2dge["ID_uniprot"].duplicated(keep=False)]

ids = data["ID_uniprot"]
data_2dge[ids.isin(ids[ids.duplicated()])]

data_2dge_final = data_2dge.drop(['Unnamed: 0'], axis=1)

## Preparation of HRMS data

# data.apply(lambda row: check_equal(row), axis=1)
K=0
for index, row in data_HRMS.iterrows() :
    ac_given = row["Accession"]
    ac_uniprot = row["AC_uniprot"]
    # double check for the non equal
    if ac_given != ac_uniprot:
#         print(index,ac_given,ac_uniprot,row["ID_uniprot"])
        ac_uniprot,id_uniprot = find_accession(row["Accession"])
#     print(ac_uniprot,id_uniprot)
        data_HRMS.loc[index, 'AC_uniprot'] = ac_uniprot
        data_HRMS.loc[index, 'ID_uniprot'] = id_uniprot
    id_uniprot = row["ID_uniprot"]
    id_given = row["Accession Name"]
    # 2 id_uniprot was equal to "in" so we fix it 
    if id_uniprot=="in":
        ac_uni,id_uni = find_accession(row["Accession"])
        data_HRMS.loc[index, 'ID_uniprot'] = id_uni
        print(id_given,ac_uniprot,id_uniprot,id_uni)


data_HRMS.to_excel("HRMS AC ID modified.xlsx")

data_HRMS_modified = pd.read_excel('HRMS AC ID modified.xlsx')

data_HRMS_modified = data_HRMS_modified.drop(['Unnamed: 0', 'Unnamed: 0.1'], axis=1)

# data.apply(lambda row: check_equal(row), axis=1)
l=0
o=0
obsolete = []
import math
for index, row in data_HRMS_modified.iterrows() :
    ac_given = row["Accession"]
    ac_uniprot = row["AC_uniprot"]
    if ac_given != ac_uniprot:
        l = l+1
#         print(index,ac_given,ac_uniprot,row["ID_uniprot"])
    if pd.isnull(ac_uniprot):
        o = o+1
        obsolete.append(index)
print(l,o)
print("Obsolete entries: "+str(o))

data_HRMS_no_obsolete = data_HRMS_modified.drop(obsolete)

data_HRMS_no_obsolete

#### Check for duplicates

duplicates = data_HRMS_no_obsolete[data_HRMS_no_obsolete["ID_uniprot"].duplicated(keep=False)]

# 12 proteins was concidered unique in midbrain but after their name being changed they are not any more
# the only duplicates are the proteins that belong only in the midbrain
duplicates

dup = []
for index, row in data_HRMS_no_obsolete.iterrows() :
    if row["ID_uniprot"] in duplicates.values:
        if row["Brain part"]=="MB":
            dup.append(index)
        else:
            data_HRMS_no_obsolete.loc[index, 'Brain part'] = data_HRMS_no_obsolete.loc[index, 'Brain part']+", MB" 

dup

data_HRMS_final= data_HRMS_no_obsolete.drop(dup)

data_HRMS_final

ids = data_HRMS_final["AC_uniprot"]
data_HRMS_final[ids.isin(ids[ids.duplicated()])]
# NO DUPLICATES

data_HRMS_final.to_excel("HRMS final.xlsx")

### data_HRMS_final data_2dge_final

These tables have ALL the correct information

d_HRMS = data_HRMS_final.drop(['Accession', 'Accession Name'], axis=1)

d_2dge = data_2dge_final.drop(['Accession Name'], axis=1)

# belong to d_HRMS and doesn't belong to d_2dge
d_HRMS_unique = d_HRMS[~d_HRMS['AC_uniprot'].isin(d_2dge ['AC_uniprot'])]
d_HRMS_unique = d_HRMS_unique.rename(columns={"Brain part": "Brain part HRMS"})
print("Unique in HRMS: " + str(len(d_HRMS_unique)))
# belong to d_2dge and doesn't belong to d_HRMS
d_2dge_unique = d_2dge[~d_2dge['AC_uniprot'].isin(d_HRMS ['AC_uniprot'])]
d_2dge_unique = d_2dge_unique.rename(columns={"Brain part": "Brain part 2dge"})
print("Unique in 2dge: " + str(len(d_2dge_unique)))
common = pd.merge(d_2dge, d_HRMS, on=["AC_uniprot","ID_uniprot"])
print("Common: " + str(len(common)))
total_prot = pd.concat([d_HRMS_unique,d_2dge_unique,common])
total_prot.to_excel("Total proteins.xlsx")

total_prot

cellular_component = pd.DataFrame(columns=['AC_uniprot', 'ID_uniprot', 'gene_id','information'])
molecular_function = pd.DataFrame(columns=['AC_uniprot', 'ID_uniprot', 'gene_id','information'])
biological_process = pd.DataFrame(columns=['AC_uniprot', 'ID_uniprot', 'gene_id','information'])

# data.apply(lambda row: check_equal(row), axis=1)
K=0
total_prot1 = total_prot.copy()
for index, row in total_prot1.iterrows() :
    ac_uniprot = row["AC_uniprot"]
    id_uniprot = row["ID_uniprot"]
    cc,mf,bp,cc_id,mf_id,bp_id = retrieve_prot_inf(ac_uniprot)
    cc_len = len(cc)
    mf_len = len(mf)
    bp_len = len(bp)
    for i in range(cc_len):
        cellular_component = cellular_component.append({'AC_uniprot': ac_uniprot,'ID_uniprot':id_uniprot,'gene_id':cc_id[i],'information':cc[i]}, ignore_index=True)
    for i in range(mf_len):
        molecular_function = molecular_function.append({'AC_uniprot': ac_uniprot,'ID_uniprot':id_uniprot,'gene_id':mf_id[i],'information':mf[i]}, ignore_index=True)
    for i in range(bp_len):
        biological_process = biological_process.append({'AC_uniprot': ac_uniprot,'ID_uniprot':id_uniprot,'gene_id':bp_id[i],'information':bp[i]}, ignore_index=True)
    


biological_process

cellular_component.to_excel("cellular component.xlsx")

molecular_function.to_excel("molecular function.xlsx")

biological_process.to_excel("biological process.xlsx")

ppppppp = pd.read_excel('biological process.xlsx')

ppppppp[ppppppp["AC_uniprot"]=="P14873"]

molecular_function[molecular_function["information"]=="actin binding"]

import re
import requests
import urllib

def word_break(text_br):
    words = re.findall(r'\b\S+\b', text_br) 
    return words

def retrieve_prot_inf(query):
    global K
    K=K+1
    print (query,"  ---  ", K, end="\r")
    ac_uniprot = ""
    id_uniprot = ""
    res = urllib.request.urlopen("http://www.uniprot.org/uniprot/" + query + ".txt").read()
    s = str(res, 'utf-8')
#     print(s)
    lines = s.splitlines()
#     print(lines)
#   print(res)
    cellular_comp = [] #C
    molecular_func = [] #F
    biological_proc = [] #P
    cellular_comp_id = [] #C
    molecular_func_id = [] #F
    biological_proc_id = [] #P
    
    for i in range(len(lines)):
        match_word=word_break(lines[i])
        if len(lines[i])>2:
            if match_word[0]=="DR" and match_word[1]=="GO":
#                 print(lines[i])
#                 print(match_word)
#                 print(len(match_word))
#                 print(len(match_word)-5)
#                 print("-------------------------------")
                gene_id = match_word[2]
                
                if match_word[3][0] == "C":
#                     print(match_word[3][2:])
                    sentence = match_word[3][2:]
                    #because we must exclude DR GO GO:342552 and the first and last word
                    for j in range(4,len(match_word)-1):
                        sentence = sentence +" "+ match_word[j]
                    cellular_comp.append(sentence)
                    cellular_comp_id.append(gene_id)
                elif match_word[3][0] == "F":
#                     print(match_word)
#                     print(match_word[3][2:])
                    sentence = match_word[3][2:]
                    #because we must exclude DR GO GO:342552 and the first and last word
                    for j in range(4,len(match_word)-1):
                        sentence = sentence +" "+ match_word[j]
                    molecular_func.append(sentence)
                    molecular_func_id.append(gene_id)
                elif match_word[3][0] == "P":
#                     print(match_word[3][2:])
                    sentence = match_word[3][2:]
                    #because we must exclude DR GO GO:342552 and the first and last word
                    for j in range(4,len(match_word)-1):
                        sentence = sentence +" "+ match_word[j]
                    biological_proc.append(sentence)
                    biological_proc_id.append(gene_id)
                
    return cellular_comp,molecular_func,biological_proc,cellular_comp_id,molecular_func_id,biological_proc_id



K=0
retrieve_prot_inf("CV028_MOUSE")
