==> Part 3.ipynb <==
import pandas as pd
import numpy as np

biological_process = pd.read_excel('biological process.xlsx')
molecular_function = pd.read_excel("molecular function.xlsx")
cellular_component = pd.read_excel("cellular component.xlsx")

biological_process = biological_process.drop(['Unnamed: 0'], axis=1)
molecular_function = molecular_function.drop(['Unnamed: 0'], axis=1)
cellular_component = cellular_component.drop(['Unnamed: 0'], axis=1)

ancestors_biological_process

ancestors_biological_process = biological_process[['gene_id', 'information']].copy()

ancestors_biological_process.drop_duplicates(subset=['gene_id'])

ancestors_biological_process

ancestors_biological_process = biological_process[['gene_id', 'information']].copy()

ancestors_biological_process = ancestors_biological_process.drop_duplicates(subset=['gene_id'])

    if gene_id != "GO:0055114" and gene_id !="GO:0097755" and gene_id !="GO:0021882" and gene_id !="GO:0010870":

from goatools.gosubdag.gosubdag import GoSubDag
# data.apply(lambda row: check_equal(row), axis=1)
K=0
for index, row in ancestors_biological_process.iterrows() :
    K=K+1
#     print (K, end="\r")
    
    gene_id = row["gene_id"]

    #     print(gene_id)
    print (gene_id," ------ ",K, end="\r")
    ancestor_1,ancestor_2,ancestor_3 = find_ancestors1(gene_id)
    ancestors_biological_process.loc[index, 'ancestor_1'] = ancestor_1
    ancestors_biological_process.loc[index, 'ancestor_2'] = ancestor_2
    ancestors_biological_process.loc[index, 'ancestor_3'] = ancestor_3
    if ancestor_1 != "" and ancestor_2 !="" and ancestor_3!= "":
        ancestors_biological_process.loc[index, 'ancestor_1_info'] = godag[ancestor_1].name
        ancestors_biological_process.loc[index, 'ancestor_2_info'] = godag[ancestor_2].name
        if ancestor_3 != '':
            ancestors_biological_process.loc[index, 'ancestor_3_info'] = godag[ancestor_3].name


ancestors_biological_process.to_excel('ancestors biological process.xlsx')

ancestors_biological_process = pd.read_excel("ancestors biological process.xlsx")

ancestors_biological_process.drop_duplicates(subset=['ancestor_2'])

ancestors_molecular_function

ancestors_molecular_function = molecular_function[['gene_id', 'information']].copy()

ancestors_molecular_function = ancestors_molecular_function.drop_duplicates(subset=['gene_id'])

from goatools.gosubdag.gosubdag import GoSubDag
# data.apply(lambda row: check_equal(row), axis=1)
K=0
for index, row in ancestors_molecular_function.iterrows() :
    K=K+1
#     print (K, end="\r")
    gene_id = row["gene_id"]
#     print(gene_id)
    print (gene_id," ------ ",K, end="\r")
    if gene_id != "GO:0030375" and gene_id !="GO:0003712" and gene_id !="GO:0045155" and gene_id != "GO:0003714" and gene_id != "GO:0030374" and gene_id != "GO:0003713":
        ancestor_1,ancestor_2,ancestor_3 = find_ancestors(gene_id)
        ancestors_molecular_function.loc[index, 'ancestor_1'] = ancestor_1
        ancestors_molecular_function.loc[index, 'ancestor_2'] = ancestor_2
        ancestors_molecular_function.loc[index, 'ancestor_3'] = ancestor_3
        ancestors_molecular_function.loc[index, 'ancestor_1_info'] = godag[ancestor_1].name
        ancestors_molecular_function.loc[index, 'ancestor_2_info'] = godag[ancestor_2].name
        if ancestor_3 != '':
            ancestors_molecular_function.loc[index, 'ancestor_3_info'] = godag[ancestor_3].name
    #     print("ssssss")


GO:0003714  ------  170
GO:0030374  ------  446
GO:0003713  ------  55
GO:0003712  ------  214

ancestors_molecular_function.to_excel('ancestors molecular function1.xlsx')

ancestors_molecular_function1 = molecular_function[['gene_id', 'information']].copy()

ancestors_molecular_function1 = ancestors_molecular_function1.drop_duplicates(subset=['gene_id'])

from goatools.gosubdag.gosubdag import GoSubDag
# data.apply(lambda row: check_equal(row), axis=1)
K=0
for index, row in ancestors_molecular_function1.iterrows() :
    K=K+1
#     print (K, end="\r")
    
    gene_id = row["gene_id"]
    if gene_id != "GO:0030375" and gene_id !="GO:0045155":
    #     print(gene_id)
        print (gene_id," ------ ",K, end="\r")
        ancestor_1,ancestor_2,ancestor_3 = find_ancestors1(gene_id)
        ancestors_molecular_function1.loc[index, 'ancestor_1'] = ancestor_1
        ancestors_molecular_function1.loc[index, 'ancestor_2'] = ancestor_2
        ancestors_molecular_function1.loc[index, 'ancestor_3'] = ancestor_3
        ancestors_molecular_function1.loc[index, 'ancestor_1_info'] = godag[ancestor_1].name
        ancestors_molecular_function1.loc[index, 'ancestor_2_info'] = godag[ancestor_2].name
        if ancestor_3 != '':
            ancestors_molecular_function1.loc[index, 'ancestor_3_info'] = godag[ancestor_3].name


ancestors_molecular_function1.to_excel('ancestors molecular function.xlsx')

ancestors_molecular_function1.drop_duplicates(subset=['ancestor_1'])

ancestors_molecular_function1.drop_duplicates(subset=['ancestor_2'])

ancestors_molecular_function1.drop_duplicates(subset=['ancestor_3'])

data1 = pd.read_excel('ancestors molecular functionnn.xlsx')
data2 = pd.read_excel('ancestors molecular function1.xlsx')

# belong to data1 and doesn't belong to data_HRMS
data1_u = data1[~data1['ancestor_1'].isin(data2 ['ancestor_1'])]
print(len(data1_u))
# belong to data_HRMS and doesn't belong to data1
data2_u = data2[~data2['ancestor_1'].isin(data1 ['ancestor_1'])]
print(len(data2_u))
common = pd.merge(data1, data2, on=["ancestor_1"])
print(len(common))

ancestors_cellular_component

ancestors_cellular_component = cellular_component[['gene_id', 'information']].copy()

ancestors_cellular_component = ancestors_cellular_component.drop_duplicates(subset=['gene_id'])

from goatools.gosubdag.gosubdag import GoSubDag
# data.apply(lambda row: check_equal(row), axis=1)
K=0
for index, row in ancestors_cellular_component.iterrows() :
    K=K+1
#     print (K, end="\r")
    gene_id = row["gene_id"]
#     print(gene_id)
    print (gene_id," ------ ",K, end="\r")
    ancestor_1,ancestor_2,ancestor_3 = find_ancestors(gene_id)
    ancestors_cellular_component.loc[index, 'ancestor_1'] = ancestor_1
    ancestors_cellular_component.loc[index, 'ancestor_2'] = ancestor_2
    ancestors_cellular_component.loc[index, 'ancestor_3'] = ancestor_3
    ancestors_cellular_component.loc[index, 'ancestor_1_info'] = godag[ancestor_1].name
    ancestors_cellular_component.loc[index, 'ancestor_2_info'] = godag[ancestor_2].name
    if ancestor_3 != '':
        ancestors_cellular_component.loc[index, 'ancestor_3_info'] = godag[ancestor_3].name
#     print("ssssss")


ancestors_cellular_component

ancestors_cellular_component.drop_duplicates(subset=['ancestor_1'])

ancestors_cellular_component.drop_duplicates(subset=['ancestor_2'])

ancestors_cellular_component.drop_duplicates(subset=['ancestor_3'])

ancestors_cellular_component.to_excel('ancestors cellular component.xlsx')

cellular_component[cellular_component["AC_uniprot"]=="P14873"]

molecular_function[molecular_function["information"]=="actin binding"]

pip install goatools

import os
from goatools.obo_parser import GODag

# Load a small test GO DAG and all the optional relationships,
# like 'regulates' and 'part_of'
godag = GODag('go-basic.obo',
              optional_attrs={'relationship'})

#Create a set
seta = set([5, 10, 3, 15, 2, 20])
#Find the length use len()
print(len(seta))

from goatools.gosubdag.gosubdag import GoSubDag
from goatools.godag.go_tasks import get_go2parents

 


def find_ancestors(GO_ID):
    optional_relationships = set()
    go2parents_isa = get_go2parents(godag, optional_relationships)

    
    not_root = True
    while(not_root):
        gosubdag_r0 = GoSubDag([GO_ID], godag, prt=None)
#         print(list(gosubdag_r0.rcntobj.go2ancestors[GO_ID]))
        if len(list(gosubdag_r0.rcntobj.go2ancestors[GO_ID]))==1:
            ancestor_1 = go2parents_isa[GO_ID]
            return (list(ancestor_1)[0],GO_ID,"")
            not_root = False 
        if len(list(gosubdag_r0.rcntobj.go2ancestors[GO_ID]))==2:
            ancestor_2 = go2parents_isa[GO_ID]
            ancestor_1 = go2parents_isa[list(ancestor_2)[0]]
            return (list(ancestor_1)[0],list(ancestor_2)[0],GO_ID)
            not_root = False
        if len(list(gosubdag_r0.rcntobj.go2ancestors[GO_ID]))==3:
            ancestor_3 = go2parents_isa[GO_ID]
#             print(ancestor_3)
            ancestor_2 = go2parents_isa[list(ancestor_3)[0]]
#             print(ancestor_2)
            ancestor_1 = go2parents_isa[list(ancestor_2)[0]]
            return (list(ancestor_1)[0],list(ancestor_2)[0],list(ancestor_3)[0])
            not_root = False
        else:
            GO_ID = list(go2parents_isa[GO_ID])[0]
        
print(find_ancestors("GO:0003714"))

from goatools.gosubdag.gosubdag import GoSubDag
from goatools.godag.go_tasks import get_go2parents

 


def find_ancestors1(GO_ID):
    optional_relationships = set()
    go2parents_isa = get_go2parents(godag, optional_relationships)
    ancestor_1 = ""
    ancestor_2 = ""
    ancestor_3 = ""
    try:
        gosubdag_r0 = GoSubDag([GO_ID], godag, prt=None)
        ancestors = list(gosubdag_r0.rcntobj.go2ancestors[GO_ID])
#         print(ancestors)
        levels = []
        for n in gosubdag_r0.rcntobj.go2ancestors[GO_ID]:
            levels.append(godag[n].level)
#         print(levels)   

    #     print(levels.index(0))
        ancestor_1 = ancestors[levels.index(0)]
        if 2 in levels:
            ancestor_3 = ancestors[levels.index(2)]
            ancestor_2 = ancestors[levels.index(1)]
        elif 1 in levels:
            ancestor_2 = ancestors[levels.index(1)]
            ancestor_3 = GO_ID
        else:
            ancestor_2 = GO_ID
            ancestor_3 = ""    
        return ancestor_1,ancestor_2,ancestor_3
    except KeyError as e:
        print ('I got a KeyError - reason' + str(e))
        return ancestor_1,ancestor_2,ancestor_3
        
print(find_ancestors1("GO:0060352"))



# Choose a deep leaf-level GO ID associated with "bacteria"
DESC = 'nucleus'            # GO Term name contains this
NSPC = 'cellular_component'  # Desired namespace

# Create a chooser function which returns True or False
def chooser(goterm):
    """Choose a leaf-level GO term based on its name"""
    b_match = DESC in goterm.name
    # True if GO term is leaf-level (has no children)
    b_leaf = not goterm.children
    # True if GO term is in 'cellular_component' namespace (nspc)
    b_nspc = goterm.namespace == NSPC
    return b_match and b_leaf and b_nspc

# Get GO terms with desired name in desired GO DAG branch
go_ids_selected = set(o.item_id for o in godag.values() if chooser(o))
print(godag.values())
print('{N} {desc} GO terms'.format(N=len(go_ids_selected), desc=DESC))

# for o in godag.values():
#     print(o)
print(godag["GO:0003714"].level)

print((godag["GO:0003674"].level))

# for n in ....:
#     if godag[n].

from goatools.gosubdag.gosubdag import GoSubDag
gosubdag = GoSubDag(go_ids_selected, godag)

go_id, go_term = max(gosubdag.go2obj.items(), key=lambda t: t[1].depth)

# Print GO ID, using print format in gosubdag
print(go_id, go_term.name)

go_ids_chosen = go_term.get_all_parents()
print('{N} ancestors for {GO} "{name}"'.format(
    N=len(go_ids_chosen), GO=go_term.item_id, name=go_term.name))

ppp = "GO:0014808"
ppp.children

godag["GO:0032991"].name


gosubdag_r0 = GoSubDag([GO_ID], godag, prt=None)
print('{GO} ancestors: {P}'.format(
    GO=GO_ID,
    P=gosubdag_r0.rcntobj.go2ancestors[GO_ID]))
