==> Part 4.ipynb <==
import pandas as pd
import numpy as np

data_sharma1 = pd.read_excel('sharma_uniprot1.xlsx')
data_sharma2 = pd.read_excel('sharma_uniprot2.xlsx')
data_jung = pd.read_excel('jung.xlsx')
data = pd.read_excel('Total proteins.xlsx')
data = data.drop(['Unnamed: 0','Brain part HRMS','Brain part 2dge','Brain part_x','Brain part_y'], axis=1)
# data_sharma = data_sharma[['Leading Protein Id','Leading Gene Name','Leading Protein Name']]
# data_sharma = data_sharma.rename(columns={"Leading Protein Id": "AC_uniprot"})
# print(data)

fields = ["Entry","Entry name","Protein names","Gene names"]
data_sharma = find_unique_common(data_sharma1,data_sharma2,"Entry",fields)

for index, row in data_sharma.iterrows() :
    data_sharma.loc[index, 'study'] = "Cell type– and brain region–resolved mouse brain proteome"
    data_sharma.loc[index, 'author'] = "Kirti Sharma et al"

for index, row in data_jung.iterrows() :
    data_jung.loc[index, 'study'] = "An Anatomically Resolved Mouse Brain Proteome Reveals Parkinson Disease-relevant Pathways"
    data_jung.loc[index, 'author'] = "Sung Yun Jung et al"

data_sharma.to_excel('sharma study.xlsx')
data_jung.to_excel('jung study.xlsx')

data_jung

sharma = pd.read_excel("sharma study.xlsx")
jung = pd.read_excel("jung study.xlsx")

sharma = sharma.drop(['Unnamed: 0','Protein names','Gene names','study','author'], axis=1)
jung = jung.drop(['Unnamed: 0','Protein names','Gene names','study','author'], axis=1)

sharma = sharma.rename(columns={"Entry": "AC_uniprot","Entry name": "ID_uniprot"})
jung = jung.rename(columns={"Entry": "AC_uniprot","Entry name": "ID_uniprot"})

sharma

#common and unique of 2 studies
fields = ["AC_uniprot","ID_uniprot"]
data_studies = find_unique_common(sharma,jung,"AC_uniprot",fields)

#common and unique of 2 studies and our study
fields = ["AC_uniprot","ID_uniprot"]
data_total = find_unique_common(data_studies,data,"AC_uniprot",fields)

ids = data_total["AC_uniprot"]
data_total[ids.isin(ids[ids.duplicated()])]
# NO DUPLICATES

data_total.to_excel('totaaaal.xlsx')

cellular_component = pd.DataFrame(columns=['AC_uniprot', 'ID_uniprot', 'gene_id','information'])
molecular_function = pd.DataFrame(columns=['AC_uniprot', 'ID_uniprot', 'gene_id','information'])
biological_process = pd.DataFrame(columns=['AC_uniprot', 'ID_uniprot', 'gene_id','information'])

data_total[14803:]

# data.apply(lambda row: check_equal(row), axis=1)
# K=0
total_prot1 = data_total.copy()
for index, row in total_prot1[14803:].iterrows() :
    ac_uniprot = row["AC_uniprot"]
    id_uniprot = row["ID_uniprot"]
    cc,mf,bp,cc_id,mf_id,bp_id = retrieve_prot_inf(ac_uniprot)
    cc_len = len(cc)
    mf_len = len(mf)
    bp_len = len(bp)
    for i in range(cc_len):
        cellular_component = cellular_component.append({'AC_uniprot': ac_uniprot,'ID_uniprot':id_uniprot,'gene_id':cc_id[i],'information':cc[i]}, ignore_index=True)
    for i in range(mf_len):
        molecular_function = molecular_function.append({'AC_uniprot': ac_uniprot,'ID_uniprot':id_uniprot,'gene_id':mf_id[i],'information':mf[i]}, ignore_index=True)
    for i in range(bp_len):
        biological_process = biological_process.append({'AC_uniprot': ac_uniprot,'ID_uniprot':id_uniprot,'gene_id':bp_id[i],'information':bp[i]}, ignore_index=True)
    


K-12

biological_process

biological_process.to_excel('biological process total.xlsx')
molecular_function.to_excel('molecular function total.xlsx')
cellular_component.to_excel('cellular component total.xlsx')

def find_unique_common(data1,data2,field,fields):
    # belong to CC_data_2dge and doesn't belong to CC_data_HRMS
    data1_unique = data1[~data1[field].isin(data2 [field])]
    print(len(data1_unique))
    # belong to CC_data_HRMS and doesn't belong to data_sharma
    data2_unique = data2[~data2[field].isin(data1 [field])]
    print(len(data2_unique))
    common = pd.merge(data1, data2, on=fields)
    print(len(common))
    total_prot = pd.concat([data1_unique,data2_unique,common])
    print(len(total_prot))
    return total_prot

import re
import requests
import urllib

def word_break(text_br):
    words = re.findall(r'\b\S+\b', text_br) 
    return words

def retrieve_prot_inf(query):
    global K
    K=K+1
    print (query,"  ---  ", K, end="\r")
    ac_uniprot = ""
    id_uniprot = ""
    res = urllib.request.urlopen("http://www.uniprot.org/uniprot/" + query + ".txt").read()
    s = str(res, 'utf-8')
#     print(s)
    lines = s.splitlines()
#     print(lines)
#   print(res)
    cellular_comp = [] #C
    molecular_func = [] #F
    biological_proc = [] #P
    cellular_comp_id = [] #C
    molecular_func_id = [] #F
    biological_proc_id = [] #P
    
    for i in range(len(lines)):
        match_word=word_break(lines[i])
        if len(lines[i])>2:
            if match_word[0]=="DR" and match_word[1]=="GO":
#                 print(lines[i])
#                 print(match_word)
#                 print(len(match_word))
#                 print(len(match_word)-5)
#                 print("-------------------------------")
                gene_id = match_word[2]
                
                if match_word[3][0] == "C":
#                     print(match_word[3][2:])
                    sentence = match_word[3][2:]
                    #because we must exclude DR GO GO:342552 and the first and last word
                    for j in range(4,len(match_word)-1):
                        sentence = sentence +" "+ match_word[j]
                    cellular_comp.append(sentence)
                    cellular_comp_id.append(gene_id)
                elif match_word[3][0] == "F":
#                     print(match_word)
#                     print(match_word[3][2:])
                    sentence = match_word[3][2:]
                    #because we must exclude DR GO GO:342552 and the first and last word
                    for j in range(4,len(match_word)-1):
                        sentence = sentence +" "+ match_word[j]
                    molecular_func.append(sentence)
                    molecular_func_id.append(gene_id)
                elif match_word[3][0] == "P":
#                     print(match_word[3][2:])
                    sentence = match_word[3][2:]
                    #because we must exclude DR GO GO:342552 and the first and last word
                    for j in range(4,len(match_word)-1):
                        sentence = sentence +" "+ match_word[j]
                    biological_proc.append(sentence)
                    biological_proc_id.append(gene_id)
                
    return cellular_comp,molecular_func,biological_proc,cellular_comp_id,molecular_func_id,biological_proc_id



pip install goatools

import os
from goatools.obo_parser import GODag

# Load a small test GO DAG and all the optional relationships,
# like 'regulates' and 'part_of'
godag = GODag('go-basic.obo',
              optional_attrs={'relationship'})

from goatools.gosubdag.gosubdag import GoSubDag
from goatools.godag.go_tasks import get_go2parents

 


def find_ancestors(GO_ID):
    optional_relationships = set()
    go2parents_isa = get_go2parents(godag, optional_relationships)
    ancestor_1 = ""
    ancestor_2 = ""
    ancestor_3 = ""
    try:
        gosubdag_r0 = GoSubDag([GO_ID], godag, prt=None)
        ancestors = list(gosubdag_r0.rcntobj.go2ancestors[GO_ID])
#         print(ancestors)
        levels = []
        for n in gosubdag_r0.rcntobj.go2ancestors[GO_ID]:
            levels.append(godag[n].level)
#         print(levels)   

    #     print(levels.index(0))
        ancestor_1 = ancestors[levels.index(0)]
        if 2 in levels:
            ancestor_3 = ancestors[levels.index(2)]
            ancestor_2 = ancestors[levels.index(1)]
        elif 1 in levels:
            ancestor_2 = ancestors[levels.index(1)]
            ancestor_3 = GO_ID
        else:
            ancestor_2 = GO_ID
            ancestor_3 = ""    
        return ancestor_1,ancestor_2,ancestor_3
    except KeyError as e:
        print ('I got a KeyError - reason' + str(e))
        return ancestor_1,ancestor_2,ancestor_3
        
print(find_ancestors("GO:0060352"))


biological_process = pd.read_excel("biological process total.xlsx")
molecular_function = pd.read_excel("molecular function total.xlsx")
cellular_component = pd.read_excel("cellular component total.xlsx")

ancestors_biological_process

ancestors_biological_process = biological_process[['gene_id', 'information']].copy()
ancestors_biological_process = ancestors_biological_process.drop_duplicates(subset=['gene_id'])

from goatools.gosubdag.gosubdag import GoSubDag
# data.apply(lambda row: check_equal(row), axis=1)
K=0
for index, row in ancestors_biological_process.iterrows() :
    K=K+1
#     print (K, end="\r")
    
    gene_id = row["gene_id"]

    #     print(gene_id)
    print (gene_id," ------ ",K, end="\r")
    ancestor_1,ancestor_2,ancestor_3 = find_ancestors(gene_id)
    ancestors_biological_process.loc[index, 'ancestor_1'] = ancestor_1
    ancestors_biological_process.loc[index, 'ancestor_2'] = ancestor_2
    ancestors_biological_process.loc[index, 'ancestor_3'] = ancestor_3
    if ancestor_1 != "" and ancestor_2 !="" and ancestor_3!= "":
        ancestors_biological_process.loc[index, 'ancestor_1_info'] = godag[ancestor_1].name
        ancestors_biological_process.loc[index, 'ancestor_2_info'] = godag[ancestor_2].name
        if ancestor_3 != '':
            ancestors_biological_process.loc[index, 'ancestor_3_info'] = godag[ancestor_3].name

ancestors_biological_process

ancestors_biological_process.to_excel('ancestors biological process total.xlsx')

data_jung = pd.read_excel('olfactory jung.xlsx')
data = pd.read_excel('Total proteins.xlsx')

data_jung = data_jung.rename(columns={"Entry": "AC_uniprot"})
ids = data_jung["Gene names"]
data_jung[ids.isin(ids[ids.duplicated()])]
# NO DUPLICATES

data_jung

fields = ["AC_uniprot"]
data_studies = find_unique_common(data_jung,data,"AC_uniprot",fields)
